chore: move label name logic to ListEnvPortBinding
